Extract GradleApiMetadata into .support internal package
